namespace API seperated Documentation
