+++ mob.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "test", "test\test.csproj", "{C8777AA4-DF09-425D-9C45-1C1327BAAB92}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Packer", "packer\Packer.csproj", "{B12F0EDB-D5E1-4340-857F-B662A48CE17C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "packer", "packer\packer.csproj", "{B12F0EDB-D5E1-4340-857F-B662A48CE17C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{50DB5190-6517-415A-A04E-5B220A8BF5EB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{B12F0EDB-D5E1-4340-857F-B662A48CE17C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B12F0EDB-D5E1-4340-857F-B662A48CE17C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B12F0EDB-D5E1-4340-857F-B662A48CE17C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50DB5190-6517-415A-A04E-5B220A8BF5EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50DB5190-6517-415A-A04E-5B220A8BF5EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50DB5190-6517-415A-A04E-5B220A8BF5EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50DB5190-6517-415A-A04E-5B220A8BF5EB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
