+++ packer/APIException.cs
@@ -1,15 +1,4 @@
-﻿namespace Packer;
-
-
-public static class Messages
-{
-    public static string MaxItemWeight = "Weight of an item excessing limits";
-    public static string MaxItemCost = "Cost of an item excessing limits";
-    public static string MaxItem = "Can not add more than maximum allowed items to the package";
-    public static string MaxPackageWeigth = "Box weight can not be more than allowed limits";
-    public static string LineParsingError = "Malformatted data within file";
-    public static string FileOpenError = "Error while trying to read file";
-}
+﻿namespace packer;
 
 public class APIException : Exception
 {
