--- packer/Conf/IConf.cs
@@ -1,10 +1,0 @@
-﻿namespace Packer
-{
-    public interface IConf
-    {
-        double MaxItemCost { get; }
-        int MaxItemCount { get; }
-        double MaxItemWeight { get; }
-        double MaxPackageWeight { get; }
-    }
-}
