+++ packer/Config/IConf.cs
@@ -1,0 +1,9 @@
+﻿namespace packer.Config;
+
+public interface IConf
+{
+    double MaxItemCost { get; }
+    int MaxItemCount { get; }
+    double MaxItemWeight { get; }
+    double MaxPackageWeight { get; }
+}
