+++ packer/Messages.cs
@@ -1,0 +1,11 @@
+﻿namespace packer;
+
+public static class Messages
+{
+    public static string MaxItemWeight = "Weight of an item excessing limits";
+    public static string MaxItemCost = "Cost of an item excessing limits";
+    public static string MaxItem = "Can not add more than maximum allowed items to the package";
+    public static string MaxPackageWeigth = "Box weight can not be more than allowed limits";
+    public static string LineParsingError = "Malformatted data within file";
+    public static string FileOpenError = "Error while trying to read file";
+}
