+++ test/test.csproj
@@ -26,7 +26,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\packer\Packer.csproj" />
+    <ProjectReference Include="..\Api\Api.csproj" />
+    <ProjectReference Include="..\packer\packer.csproj" />
   </ItemGroup>
 
 </Project>
